+++ .agentplane/tasks/202609121424-T83XJA/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202609121424-T83XJA/implement-durable-0-7-9-usage-cost-and-latency-a",
   "created_at": "2026-09-12T20:52:29.086Z",
-  "diffstat_sha256": "sha256:3eb7dbca6b74b8716a633e6a4c386d64a0fc4f8d59af8cb80317aba152a1077b",
+  "diffstat_sha256": "sha256:a0bf04ad658452280cec8ecf506df7d32b4f4e3c8fda72399a6c3caff5338d06",
   "last_verified_at": null,
   "provider": {
     "hostname": "github.com",
