+++ packages/agentplane/src/commands/task/external-agent-supervisor.ts
@@ -1,5 +1,4 @@
 import { captureExternalTaskArtifacts } from "./external-agent-task-artifact-baseline.js";
-import { createHash } from "node:crypto";
 import { readFile } from "node:fs/promises";
 import path from "node:path";
 
@@ -24,10 +23,10 @@ import {
   resolveCommandGitCommonDir,
   type CommandContext,
 } from "../shared/task-backend.js";
-
 import { agentTransitionId } from "./agent-action-packet.js";
 import {
   externalAgentIssueDigest,
+  externalAgentExchangeDigest,
   externalAgentResultDigest,
   externalAgentUsageAccounting,
   persistExternalAgentExchangeArtifacts,
@@ -64,16 +63,11 @@ import {
 } from "./external-agent-result-routing.js";
 import { readDirectRepositoryStatus, readDirectTaskHead } from "./direct-task-finalization.js";
 import { resolveConflictReworkSemanticInput } from "../pr/conflict-rework-semantic-input.js";
-
 export type IssuedExternalAgentExchange = {
   exchange: ExternalAgentExchange;
   paths: ExternalAgentExchangePaths;
   work_order: AgentWorkOrderV2;
 };
-
-function digestText(value: string): string {
-  return `sha256:${createHash("sha256").update(value, "utf8").digest("hex")}`;
-}
 
 async function commandContextForCheckout(opts: {
   command: CommandContext;
@@ -126,7 +120,7 @@ async function prepareEvaluatorInput(opts: {
       work_order: opts.work_order,
       git_root: packet.git_root,
       work_order_path: prepared.work_order_path,
-      digest: digestText(serialized),
+      digest: externalAgentExchangeDigest(serialized),
     }),
     evaluator_work_order_ref: prepared.work_order_path,
   };
@@ -327,7 +321,9 @@ async function assertReadOnlyReturnFresh(opts: {
     );
     if (
       !opts.exchange.evaluator_work_order_ref ||
-      digestText(await readFile(opts.exchange.evaluator_work_order_ref, "utf8")) !== frozen?.digest
+      externalAgentExchangeDigest(
+        await readFile(opts.exchange.evaluator_work_order_ref, "utf8"),
+      ) !== frozen?.digest
     ) {
       throw new CliError({
         code: "E_VALIDATION",
